+++ glean/src/cli.ts
@@ -18,7 +18,7 @@ const execAsync = promisify(exec.exec);
 const VIRTUAL_ENVIRONMENT_DIR = ".venv";
 
 // The version of glean_parser to install from PyPI.
-const GLEAN_PARSER_VERSION = "2.5.0";
+const GLEAN_PARSER_VERSION = "3.1.0";
 
 // This script runs a given Python module as a "main" module, like
 // `python -m module`. However, it first checks that the installed